% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_classifier.R
\name{apply_classifier}
\alias{apply_classifier}
\title{Apply a classifier induced with function build_classifier}
\usage{
apply_classifier(model, X, only.best.classifiers = FALSE)
}
\arguments{
\item{model}{model built using function build_classifier}

\item{X}{matrix defining the input space of your test set}

\item{only.best.classifiers}{if TRUE, only the most performing classification functions will be considered}
}
\value{
prediction results
}
\description{
This function applies the set of SVM classifiers to perform the supervised learning task
based on the topological data analysis
}
\section{References}{
 de Mello, R.F. (2019) "On the Shattering Coefficient of Supervised Learning Algorithms" arXiv:\url{https://arxiv.org/abs/1911.05461}

 de Mello, R.F., Ponti, M.A. (2018, ISBN: 978-3319949888) "Machine Learning: A Practical Approach on the Statistical Learning Theory"
}

\keyword{analysis}
\keyword{coefficient}
\keyword{complexity}
\keyword{dataset}
\keyword{for}
\keyword{of}
\keyword{shattering}
\keyword{some}
\keyword{the}
