% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_space.R
\name{compress_space}
\alias{compress_space}
\title{Function to compress the space given the equivalence relations.}
\usage{
compress_space(M, Y)
}
\arguments{
\item{M}{sparse matrix representing all equivalence relations}

\item{Y}{numerical vector indentifying the output space of variables}
}
\value{
A list containing sparse vectors (from package Matrix) identifying the equivalence relations
}
\description{
This function compresses the input space according to the equivalence relations, i.e., it compresses whenever an example has other elements inside its open ball but having the same class label as the ball-centered instance.
}
\keyword{compress}
\keyword{space}
