% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalence_relation.R
\name{equivalence_relation}
\alias{equivalence_relation}
\title{Function to assess all equivalence relations.}
\usage{
equivalence_relation(X, Y, quantile.percentage = 0.05, epsilon = 1e-07)
}
\arguments{
\item{X}{matrix indentifying the input space of variables}

\item{Y}{numerical vector indentifying the output space of variables}

\item{quantile.percentage}{real number to define the quantile of distances to be considered (e.g. 0.1 means 10\%)}

\item{epsilon}{a real threshold to be removed from distances in order to measure the open balls in the underlying topology}
}
\value{
A list with the equivalence relations in form of a list
}
\description{
This function computes the greatest as possible open ball connecting a given input example to every other under the same class label, thus homogeneizing space regions.
}
\keyword{equivalence}
\keyword{relation}
